+++ src/$zip-all.d.ts
@@ -1,5 +1,10 @@
 import { $InputIterable, $GeneratorIterator } from './internal/$iterable';
 
+declare function $zipAll<F, T = any>(
+  options: { filler?: F },
+  ...iterables: Array<$InputIterable<T>>
+): $GeneratorIterator<Array<T | F>>;
+
 declare function $zipAll<T = any>(
   ...iterables: Array<$InputIterable<T>>
 ): $GeneratorIterator<Array<T | undefined>>;