+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Optional configuration arguments can no longer be undefined. This was at odds with considering undefined as a valid iterable.
  - find takes an extra optional argument: the value to return if the nothing is found.
  - range can now called as either `range(end)` or `range(start, end, step)`. This matches Python.
+ - zipAll now takes optional `filler` argument to use in place of values from exhausted iterables.
  - fork/asyncFork takes an extra optional argument: the number of forks.
  - permutations, combinations, combinationsWithReplacement: order of arguments is changed. Can now be curried.
  - groupBy()/asyncGroupBy(null) now return a partial application instead of an iterable.